// src/app/modules/kentico/models/landing-bike.model.ts
import { ContentItem, Elements } from 'kentico-cloud-delivery';
export class LandingBike extends ContentItem {

  title: Elements.RichTextElement;
  background_immage: Elements.RichTextElement;
  subtitle: Elements.RichTextElement;
  subsectiontitle: Elements.RichTextElement;
  title_card: Elements.LinkedItemsElement;
  image_card: Elements.LinkedItemsElement;
}